+++ svelte/src/lib/CategoryMap.svelte
@@ -1,11 +1,10 @@
-<script>
+<script lang="ts">
 	import 'leaflet/dist/leaflet.css';
 	import { LeafletMap } from '../services/leaflet-map';
 	import { onMount } from 'svelte';
-	// @ts-ignore
-	import { latestPOI, user } from '../stores';
+
 	import { geotagService } from '../services/geotag-service';
-    let userPoi = [];
+	let userPoi = [];
 	let userCategories = [];
 	const geotagCredentials = localStorage.geotag;
 	const userData = JSON.parse(geotagCredentials);
@@ -18,41 +17,39 @@
 	};
 
 	export let mapSize = '500';
-	/**
-	 * @type {LeafletMap | null}
-	 */
-	let map = null;
+
+	let map: LeafletMap | null = null;
 
 	onMount(async () => {
 		// get the list of categories
-        userCategories = await geotagService.getUserCategories(userID);
-         // make a map
+		userCategories = await geotagService.getUserCategories(userID);
+		// make a map
 		map = new LeafletMap('category-map', mapConfig);
 		map.showZoomControl();
-        
+
 		// for each category add the points of interest
-        // @ts-ignore
-        userCategories.forEach(async (category)=>{
-           // add the layer
-           // @ts-ignore
-           map.addLayerGroup(category.title); 
-           // add the pois
-           userPoi = await geotagService.getPoiList(category._id);     
-           // for each point add map location
-           // @ts-ignore
-           userPoi.forEach((poi) =>{
-            addPoiMarker(map, poi,category.title);
-           });
-        });
-        map.showLayerControl();
+
+		userCategories.forEach(async (category: any) => {
+			// add the layer
+
+			map.addLayerGroup(category.title);
+			// add the poi
+			userPoi = await geotagService.getPoiList(category._id);
+			// for each point add map location
+
+			userPoi.forEach((poi:any) => {
+				addPoiMarker(map, poi, category.title);
+			});
+		});
+		map.showLayerControl();
 	});
 
-	// @ts-ignore
-	function addPoiMarker(map, in_Poi,in_Category) {
-		const poiStr = `${in_Poi.name}`;
-		map.addMarker({ lat: in_Poi.latitude, lng: in_Poi.longitude }, poiStr, in_Category);
+	function addPoiMarker(map: LeafletMap | null, in_Poi: any, in_Category: any) {
+		var popupString = `<a href='/dashboard/poi/weather/?_id=${in_Poi._id}'>${in_Poi.name}<br><small>Click for Weather</small></a>`;
+		// var popupString = `${in_Poi.name}`;
+		map.addMarker({ lat: in_Poi.latitude, lng: in_Poi.longitude }, popupString, in_Category);
 	}
-   // toDo : add subscription for Live update
+	// toDo : add subscription for Live update
 </script>
 
 <div class="box" id="category-map" style="height:{mapSize}px" />